refactor(ui): move overlays and variable-completion into subdirectories
