+++ src/ui/index.ts
@@ -24,10 +24,10 @@ export { getHelpSections } from "./helpTexts"
 export type { HelpKey, HelpSection } from "./helpTexts"
 export { ThemeProvider, useTheme, ThemePickerOverlay } from "./theme"
 export type { Theme } from "./theme"
-export { CollectionSwitcherOverlay } from "./CollectionSwitcherOverlay"
+export { CollectionSwitcherOverlay } from "./overlays/CollectionSwitcherOverlay"
 export { THEMES, contrastOnPrimary } from "./theme"
-export { PickerOverlay } from "./PickerOverlay"
-export type { PickerOverlayProps } from "./PickerOverlay"
+export { PickerOverlay } from "./overlays/PickerOverlay"
+export type { PickerOverlayProps } from "./overlays/PickerOverlay"
 export { PaneBorder, FullBorder, LeftBar } from "./borders"
 export { Select } from "./Select"
 export type { SelectItem, SelectProps } from "./Select"
