+++ src/ui/overlays/HelpOverlay.tsx
@@ -1,7 +1,7 @@
-import { getHelpSections } from "./helpTexts"
-import { useTheme } from "./theme"
+import { getHelpSections } from "../helpTexts"
+import { useTheme } from "../theme"
 import { Overlay } from "./Overlay"
-import type { Keybinds } from "./keybind"
+import type { Keybinds } from "../keybind"
 import { TextAttributes } from "@opentui/core"
 import type { ScrollBoxRenderable } from "@opentui/core"
 import { useRef } from "react"
