+++ src/ui/overlays/NewFolderOverlay.tsx
@@ -7,38 +7,37 @@ import {
 } from "react"
 import type { InputRenderable } from "@opentui/core"
 import { Overlay } from "./Overlay"
-import { useTheme } from "./theme"
+import { useTheme } from "../theme"
 
-export interface CloneRequestOverlayHandle {
+export interface NewFolderOverlayHandle {
   confirm: () => string | null
 }
 
-interface CloneRequestOverlayProps {
+interface NewFolderOverlayProps {
   visible: boolean
-  initialName: string
 }
 
-export const CloneRequestOverlay = forwardRef<
-  CloneRequestOverlayHandle,
-  CloneRequestOverlayProps
->(function CloneRequestOverlay({ visible, initialName }, ref) {
+export const NewFolderOverlay = forwardRef<
+  NewFolderOverlayHandle,
+  NewFolderOverlayProps
+>(function NewFolderOverlay({ visible }, ref) {
   const theme = useTheme()
-  const [name, setName] = useState(initialName)
+  const [name, setName] = useState("")
   const [errorText, setErrorText] = useState<string | null>(null)
   const nameRef = useRef<InputRenderable | null>(null)
 
   useEffect(() => {
     if (visible) {
-      setName(initialName)
+      setName("")
       setErrorText(null)
       nameRef.current?.focus()
     }
-  }, [visible, initialName])
+  }, [visible])
 
   useImperativeHandle(ref, () => ({
     confirm: () => {
       if (name.trim() === "") {
-        setErrorText("Request name is required")
+        setErrorText("Folder name is required")
         return null
       }
       setErrorText(null)
@@ -56,16 +55,16 @@ export const CloneRequestOverlay = forwardRef<
           paddingX: 2,
         }}
       >
-        <text fg={theme.text}>Clone Request</text>
+        <text fg={theme.text}>New Folder</text>
         <text fg={theme.textMuted}>esc</text>
       </box>
 
       <box style={{ paddingX: 2, flexDirection: "column", paddingBottom: 1 }}>
-        <text fg={theme.textMuted}>Request Name</text>
+        <text fg={theme.textMuted}>Folder Name</text>
         <input
           ref={nameRef}
           value={name}
-          placeholder="Request Name"
+          placeholder="Folder Name"
           onInput={setName}
           focused
           backgroundColor={theme.backgroundElement}
@@ -90,10 +89,10 @@ export const CloneRequestOverlay = forwardRef<
           paddingX: 2,
         }}
       >
-        <text fg={theme.text}>^S</text>
+        <text fg={theme.primary}>^S</text>
         <text fg={theme.textMuted}>save</text>
         <text fg={theme.textMuted}> · </text>
-        <text fg={theme.text}>esc</text>
+        <text fg={theme.primary}>esc</text>
         <text fg={theme.textMuted}>close</text>
       </box>
     </Overlay>
