+++ src/ui/useOverlayIntercepts.ts
@@ -5,9 +5,9 @@ import type { SaveState } from "./saveState"
 import type { Focus } from "./focus"
 import type { UseEnvironmentEditorResult } from "../hooks/useEnvironmentEditor"
 import type { EnvHeaderPaneHandle } from "./EnvHeaderPane"
-import type { NewRequestOverlayHandle } from "./NewRequestOverlay"
-import type { CloneRequestOverlayHandle } from "./CloneRequestOverlay"
-import type { NewFolderOverlayHandle } from "./NewFolderOverlay"
+import type { NewRequestOverlayHandle } from "./overlays/NewRequestOverlay"
+import type { CloneRequestOverlayHandle } from "./overlays/CloneRequestOverlay"
+import type { NewFolderOverlayHandle } from "./overlays/NewFolderOverlay"
 import type { UseRequestDraftResult } from "../hooks/useRequestDraft"
 import type { UseFolderDraftResult } from "../hooks/useFolderDraft"
