+++ src/ui/variable-completion/highlightOffsets.ts
