+++ src/ui/variable-completion/useVariableCompletion.ts
