+++ src/ui/variable-completion/variableCompletionInterceptor.tsx
