+++ tests/unit/CodeEditorCompletion.test.tsx
@@ -8,7 +8,7 @@ import { KeymapProvider } from "@opentui/keymap/react"
 import type { KeymapProviderProps } from "@opentui/keymap/react"
 import { CodeEditorRenderable } from "../../src/ui/editor/CodeEditor"
 import { CodeEditorCompletion } from "../../src/ui/editor/CodeEditorCompletion"
-import { VariableCompletionInterceptor } from "../../src/ui/variableCompletionInterceptor"
+import { VariableCompletionInterceptor } from "../../src/ui/variable-completion/variableCompletionInterceptor"
 import { ThemeProvider } from "../../src/ui/theme"
 import type { Environment } from "../../src/schema"
 import { opencodeTheme } from "../../src/ui/theme-data"
