+++ tests/unit/DeleteConfirmation.test.tsx
@@ -8,7 +8,7 @@ import {
 import { KeymapProvider } from "@opentui/keymap/react"
 import type { KeymapProviderProps } from "@opentui/keymap/react"
 import { ThemeProvider } from "../../src/ui/theme"
-import { ConfirmOverlay } from "../../src/ui/ConfirmOverlay"
+import { ConfirmOverlay } from "../../src/ui/overlays/ConfirmOverlay"
 
 function setupKeymap() {
   const { keymap, cleanup: hostCleanup } = createTestKeymap()
